# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Billing Service API",
    "version": "0.1.0",
    "description": "Manages billing accounts, credit balance, deductions, and Stripe integration for the Conductor platform."
  },
  "servers": [
    {
      "url": "http://localhost:3012"
    }
  ],
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "AuthorizeResponse": {
        "type": "object",
        "properties": {
          "sufficient": {
            "type": "boolean"
          },
          "balance_cents": {
            "type": "string"
          },
          "required_cents": {
            "type": "string"
          }
        },
        "required": [
          "sufficient",
          "balance_cents",
          "required_cents"
        ]
      },
      "AuthorizeCostItem": {
        "type": "object",
        "properties": {
          "costName": {
            "type": "string",
            "minLength": 1
          },
          "quantity": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true
          }
        },
        "required": [
          "costName",
          "quantity"
        ]
      },
      "AuthorizeRequest": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuthorizeCostItem"
            },
            "minItems": 1
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "items"
        ]
      },
      "UsageApplyResponse": {
        "type": "object",
        "properties": {
          "acknowledged": {
            "type": "boolean"
          },
          "topup_triggered": {
            "type": "boolean"
          }
        },
        "required": [
          "acknowledged",
          "topup_triggered"
        ]
      },
      "UsageApplyRequest": {
        "type": "object",
        "properties": {
          "spent_total_cents": {
            "type": "string"
          }
        },
        "required": [
          "spent_total_cents"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/v1/customer_balance/authorize": {
      "post": {
        "summary": "Synchronous pre-execution authorization with auto-topup",
        "description": "Resolves prices from costs-service, fetches usage from runs-service, fetches paid balance from stripe-service, and composes with local promo credits. If insufficient and auto-topup is configured, calls stripe-service reload (synchronous, with per-org coalescing).",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "x-api-key",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "x-user-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "x-run-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Campaign ID injected by workflow-service"
            },
            "required": false,
            "name": "x-campaign-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Brand ID(s) injected by workflow-service (comma-separated UUIDs for multi-brand campaigns)",
              "example": "uuid1,uuid2,uuid3"
            },
            "required": false,
            "name": "x-brand-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Workflow slug injected by workflow-service"
            },
            "required": false,
            "name": "x-workflow-slug",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Feature slug for tracking"
            },
            "required": false,
            "name": "x-feature-slug",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Audience ID injected by campaign-service for per-audience cost attribution"
            },
            "required": false,
            "name": "x-audience-id",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorizeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authorization result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorizeResponse"
                }
              }
            }
          },
          "502": {
            "description": "Downstream service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/customer_balance/usage_apply": {
      "post": {
        "summary": "Notify billing of an org's current usage total (hint for proactive topup)",
        "description": "Fire-and-forget endpoint called by runs-service after every runs_costs write. Billing computes balance = stripe paid topups + local credits − usage; if below topup_threshold and auto-topup is configured, fires a stripe-service reload. Always returns 202.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "x-api-key",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "x-user-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "x-run-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Campaign ID injected by workflow-service"
            },
            "required": false,
            "name": "x-campaign-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Brand ID(s) injected by workflow-service (comma-separated UUIDs for multi-brand campaigns)",
              "example": "uuid1,uuid2,uuid3"
            },
            "required": false,
            "name": "x-brand-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Workflow slug injected by workflow-service"
            },
            "required": false,
            "name": "x-workflow-slug",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Feature slug for tracking"
            },
            "required": false,
            "name": "x-feature-slug",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Audience ID injected by campaign-service for per-audience cost attribution"
            },
            "required": false,
            "name": "x-audience-id",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsageApplyRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Notification acknowledged",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageApplyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}